# api/admin/center_api.http.go
// Code generated by xapi. DO NOT EDIT.

package admin

import "github.com/gin-gonic/gin"

type centerApiServer struct {
	svc CenterApiServer
}

func (o *centerApiServer) List(c *gin.Context) (interface{}, error) {
	req := new(CenterApiListRequest)

	if err := c.ShouldBind(req); err != nil {
		return nil, err
	}

	resp, err := o.svc.List(c.Request.Context(), req)
	if err != nil {
		return nil, err
	}
	return resp, nil
}

func (o *centerApiServer) All(c *gin.Context) (interface{}, error) {
	req := new(CenterApiAllRequest)

	if err := c.ShouldBind(req); err != nil {
		return nil, err
	}

	resp, err := o.svc.All(c.Request.Context(), req)
	if err != nil {
		return nil, err
	}
	return resp, nil
}

func (o *centerApiServer) Create(c *gin.Context) (interface{}, error) {
	req := new(CenterApiCreateRequest)

	if err := c.ShouldBind(req); err != nil {
		return nil, err
	}

	resp, err := o.svc.Create(c.Request.Context(), req)
	if err != nil {
		return nil, err
	}
	return resp, nil
}

func (o *centerApiServer) Update(c *gin.Context) (interface{}, error) {
	req := new(CenterApiUpdateRequest)

	if err := c.ShouldBind(req); err != nil {
		return nil, err
	}

	resp, err := o.svc.Update(c.Request.Context(), req)
	if err != nil {
		return nil, err
	}
	return resp, nil
}

func (o *centerApiServer) SwitchStatus(c *gin.Context) (interface{}, error) {
	req := new(CenterApiSwitchStatusRequest)

	if err := c.ShouldBind(req); err != nil {
		return nil, err
	}

	resp, err := o.svc.SwitchStatus(c.Request.Context(), req)
	if err != nil {
		return nil, err
	}
	return resp, nil
}

func RegisterCenterApiServer(r *gin.Engine, svc CenterApiServer, handle func(func(c *gin.Context) (interface{}, error)) gin.HandlerFunc, middlewares ...gin.HandlerFunc) {
	server := &centerApiServer{
		svc: svc,
	}

	r.POST("/admin/center/api/list", append(middlewares, handle(server.List))...)
	r.POST("/admin/center/api/all", append(middlewares, handle(server.All))...)
	r.POST("/admin/center/api/create", append(middlewares, handle(server.Create))...)
	r.POST("/admin/center/api/update", append(middlewares, handle(server.Update))...)
	r.POST("/admin/center/api/switch/status", append(middlewares, handle(server.SwitchStatus))...)
}