# api/admin/openapi.yml
openapi: 3.0.0
info:
    title: admin api
    version: 1.0.0
...
paths:
...
    /admin/auth/captcha:
        post:
            description: 验证码
            operationId: AuthServer_Captcha
            parameters:
                - description: Accept Language
                  in: header
                  name: Accept-Language
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthCaptchaRequest'
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthCaptchaResponse'
                    description: OK
            summary: 验证码
            tags:
                - AuthServer
    /admin/auth/login:
        post:
            description: 用户登陆
            operationId: AuthServer_AuthLogin
            parameters:
                - description: Accept Language
                  in: header
                  name: Accept-Language
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthLoginRequest'
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthLoginResponse'
                    description: OK
            summary: 用户登陆
            tags:
                - AuthServer
    /admin/auth/menus:
        post:
            description: 用户菜单
            operationId: AuthServer_AuthMenus
            parameters:
                - description: Accept Language
                  in: header
                  name: Accept-Language
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthMenusRequest'
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AuthMenusResponse'
                    description: OK
            summary: 用户菜单
            tags:
                - AuthServer
    /admin/center/api/all:
        post:
            description: 后台api列表
            operationId: CenterApiServer_All
            parameters:
                - description: Accept Language
                  in: header
                  name: Accept-Language
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CenterApiAllRequest'
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CenterApiAllResponse'
                    description: OK
            summary: 后台api列表
            tags:
                - CenterApiServer
    /admin/center/api/create:
        post:
            description: 创建后台api
            operationId: CenterApiServer_Create
            parameters:
                - description: Accept Language
                  in: header
                  name: Accept-Language
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CenterApiCreateRequest'
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CenterApiCreateResponse'
                    description: OK
            summary: 创建后台api
            tags:
                - CenterApiServer
    /admin/center/api/list:
        post:
            description: 后台api列表
            operationId: CenterApiServer_List
            parameters:
                - description: Accept Language
                  in: header
                  name: Accept-Language
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CenterApiListRequest'
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CenterApiListResponse'
                    description: OK
            summary: 后台api列表
            tags:
                - CenterApiServer
    /admin/center/api/switch/status:
        post:
            description: 切换后台api状态
            operationId: CenterApiServer_SwitchStatus
            parameters:
                - description: Accept Language
                  in: header
                  name: Accept-Language
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CenterApiSwitchStatusRequest'
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CenterApiSwitchStatusResponse'
                    description: OK
            summary: 切换后台api状态
            tags:
                - CenterApiServer
    /admin/center/api/update:
        post:
            description: 更新后台api
            operationId: CenterApiServer_Update
            parameters:
                - description: Accept Language
                  in: header
                  name: Accept-Language
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CenterApiUpdateRequest'
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CenterApiUpdateResponse'
                    description: OK
            summary: 更新后台api
            tags:
                - CenterApiServer
...
    /admin/upload/image:
        post:
            description: 图片上传
            parameters:
                - description: Authorization token
                  in: header
                  name: Authorization
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            properties:
                                file:
                                    format: binary
                                    type: string
                            type: object
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                properties:
                                    url:
                                        description: 图片地址
                                        type: string
                                type: object
                    description: OK
            summary: 图片上传
            tags:
                - UploadServer
...
components:
    schemas:
...
        AuthCaptchaRequest:
            properties: null
            type: object
        AuthCaptchaResponse:
            properties:
                captcha_id:
                    description: 验证码ID
                    type: string
                captcha_image:
                    description: 验证码图片
                    type: string
            type: object
        AuthLoginRequest:
            properties:
                captcha_id:
                    description: 验证码ID
                    type: string
                password:
                    description: 密码
                    type: string
                username:
                    description: 用户名
                    type: string
                verify_code:
                    description: 验证码
                    type: string
            type: object
        AuthLoginResponse:
            properties:
                nickname:
                    description: 昵称
                    type: string
                token:
                    description: token
                    type: string
                user_id:
                    description: 用户id
                    type: string
                username:
                    description: 用户名
                    type: string
            type: object
        AuthMenusRequest:
            properties: null
            type: object
        AuthMenusResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/CenterMenuInfo'
                    type: array
            type: object
...
        CenterApiAllItem:
            properties:
                id:
                    description: ""
                    type: integer
                name:
                    description: api名称
                    type: string
                path:
                    description: api路径
                    type: string
            type: object
        CenterApiAllRequest:
            properties:
                status:
                    description: 状态
                    type: string
            type: object
        CenterApiAllResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/CenterApiAllItem'
                    type: array
            type: object
        CenterApiCreateRequest:
            properties:
                name:
                    description: api名称
                    type: string
                path:
                    description: api路径
                    type: string
            type: object
        CenterApiCreateResponse:
            properties: null
            type: object
        CenterApiInfo:
            properties:
                created_at:
                    description: 创建时间
                    type: string
                id:
                    description: ""
                    type: integer
                name:
                    description: api名称
                    type: string
                path:
                    description: api路径
                    type: string
                status:
                    description: 状态
                    type: string
            type: object
        CenterApiListRequest:
            properties:
                name:
                    description: api名称
                    type: string
                page:
                    description: ""
                    type: integer
            type: object
        CenterApiListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/CenterApiInfo'
                    type: array
                total:
                    description: ""
                    type: integer
            type: object
        CenterApiSwitchStatusRequest:
            properties:
                id:
                    description: ""
                    type: integer
                status:
                    description: ""
                    type: string
            type: object
        CenterApiSwitchStatusResponse:
            properties: null
            type: object
        CenterApiUpdateRequest:
            properties:
                id:
                    description: ""
                    type: integer
                name:
                    description: api名称
                    type: string
                path:
                    description: api路径
                    type: string
            type: object
        CenterApiUpdateResponse:
            properties: null
            type: object
...
        CenterMenuInfo:
            properties:
                created_at:
                    description: ""
                    type: string
                icon:
                    description: ""
                    type: string
                id:
                    description: ""
                    type: integer
                name:
                    description: ""
                    type: string
                name_json:
                    $ref: '#/components/schemas/model.Lang'
                path:
                    description: ""
                    type: string
                pid:
                    description: ""
                    type: integer
                sort_num:
                    description: ""
                    type: integer
                status:
                    description: ""
                    type: string
                updated_at:
                    description: ""
                    type: string
            type: object
...
        model.Lang:
            properties:
                en:
                    description: english value
                    type: string
                zh:
                    description: chinese value
                    type: string
            type: object
...